Fix balance param

+++ oauth-rest-api-server/src/main/resources/schema.sql
@@ -3,7 +3,7 @@ create table transfers (
   unique_id VARCHAR(255) PRIMARY KEY,
   transaction_id VARCHAR(255),
   usage VARCHAR(255),
-  amount DECIMAL(255),
+  amount REAL(10),
   currency VARCHAR(255),
   source_wallet_id VARCHAR(255),
   return_success_url VARCHAR(255),
